Slett mer ubrukt kode og flytt MinimertUregistrertBarn til BrevPeriodeUtil.kt
